% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseAverageQualityPlot.R
\name{BaseAverageQualityPlot}
\alias{BaseAverageQualityPlot}
\title{BaseAverageQualityPlot}
\usage{
BaseAverageQualityPlot(input_data)
}
\arguments{
\item{input_data}{folder containing data}
}
\description{
BaseAverageQualityPlot
}
