% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseAverageQualityPlotly.R
\name{BaseAverageQualityPlotly}
\alias{BaseAverageQualityPlotly}
\title{interactive BaseAverageQualityPlot}
\usage{
BaseAverageQualityPlotly(input_data)
}
\arguments{
\item{input_data}{folder containing data}
}
\description{
interactive BaseAverageQualityPlot
}
