% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseCompositionAreaChartPlot.R
\name{BaseCompositionAreaChartPlot}
\alias{BaseCompositionAreaChartPlot}
\title{BaseCompositionAreaChartPlot}
\usage{
BaseCompositionAreaChartPlot(input_data)
}
\arguments{
\item{input_data}{folder containing data}
}
\description{
BaseCompositionAreaChartPlot
}
