% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseCompositionLinePlot.R
\name{BaseCompositionLinePlot}
\alias{BaseCompositionLinePlot}
\title{BaseCompositionLinePlot}
\usage{
BaseCompositionLinePlot(input_data)
}
\arguments{
\item{input_data}{folder containing data}
}
\description{
BaseCompositionLinePlot
}
