% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseQualityBoxplotPlot.R
\name{BaseQualityBoxplotPlot}
\alias{BaseQualityBoxplotPlot}
\title{BaseQualityBoxplotPlot}
\usage{
BaseQualityBoxplotPlot(input_data)
}
\arguments{
\item{input_data}{folder containing data}
}
\description{
BaseQualityBoxplotPlot
}
