% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsServerLogic.R
\name{DEGsServerLogic}
\alias{DEGsServerLogic}
\title{Server function for DEGs module in Shiny application}
\usage{
DEGsServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Server function for DEGs module in Shiny application
}
