% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsUserInterface.R
\name{DEGsUserInterface}
\alias{DEGsUserInterface}
\title{UI function for DEGs module in Shiny application}
\usage{
DEGsUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
UI function for DEGs module in Shiny application
}
