% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAServerLogic.R
\name{EDAServerLogic}
\alias{EDAServerLogic}
\title{Server function for EDA module in Shiny application}
\usage{
EDAServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Server function for EDA module in Shiny application
}
