% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAUserInterface.R
\name{EDAUserInterface}
\alias{EDAUserInterface}
\title{UI function for EDA module in Shiny application}
\usage{
EDAUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
UI function for EDA module in Shiny application
}
