% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteringServerLogic.R
\name{FilteringServerLogic}
\alias{FilteringServerLogic}
\title{Server function for filtering module in Shiny application}
\usage{
FilteringServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Server function for filtering module in Shiny application
}
