% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteringUserInterface.R
\name{FilteringUserInterface}
\alias{FilteringUserInterface}
\title{UI function for filtering module in Shiny application}
\usage{
FilteringUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
UI function for filtering module in Shiny application
}
