% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontentDistributionPlot.R
\name{GCcontentDistributionPlot}
\alias{GCcontentDistributionPlot}
\title{GCcontentDistributionPlot}
\usage{
GCcontentDistributionPlot(input_data)
}
\arguments{
\item{input_data}{samples folder}
}
\description{
GCcontentDistributionPlot
}
