% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontentDistributionPlotly.R
\name{GCcontentDistributionPlotly}
\alias{GCcontentDistributionPlotly}
\title{interactive GCcontentDistributionPlot}
\usage{
GCcontentDistributionPlotly(input_data)
}
\arguments{
\item{input_data}{samples folder}
}
\description{
interactive GCcontentDistributionPlot
}
