% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{inDAGO}
\alias{inDAGO}
\title{inDAGO}
\usage{
inDAGO()
}
\value{
No return value, called for side effects
}
\description{
A Shiny app for dual and bulk RNA‑sequencing analysis.
}
\details{
This function allows to launch inDAGO Shiny interface.
}
