% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingBulkServerLogic.R
\name{IndexingBulkServerLogic}
\alias{IndexingBulkServerLogic}
\title{Indexing bulk server logic}
\usage{
IndexingBulkServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing bulk server logic
}
