% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingBulkUserInterface.R
\name{IndexingBulkUserInterface}
\alias{IndexingBulkUserInterface}
\title{Indexing bulk ui}
\usage{
IndexingBulkUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing bulk ui
}
