% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingComb.R
\name{IndexingComb}
\alias{IndexingComb}
\title{Combined indexing}
\usage{
IndexingComb(
  basename,
  reference,
  gappedIndex,
  indexSplit,
  memory,
  TH_subread,
  gen1,
  gen2,
  outfolder,
  tempfolder = file.path(fs::path_temp(), "TempDirSum_3738"),
  tag1,
  tag2
)
}
\arguments{
\item{basename}{output basename}

\item{reference}{reference genome}

\item{gappedIndex}{gapped structure}

\item{indexSplit}{split structure}

\item{memory}{handling memory}

\item{TH_subread}{threshold memory usage}

\item{gen1}{first reference genome}

\item{gen2}{second reference genome}

\item{outfolder}{output folder}

\item{tempfolder}{temporary folder}

\item{tag1}{first genome label}

\item{tag2}{second genome label}
}
\description{
Combined indexing
}
