% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingCombinedServerLogic.R
\name{IndexingCombinedServerLogic}
\alias{IndexingCombinedServerLogic}
\title{Indexing combined server logic}
\usage{
IndexingCombinedServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing combined server logic
}
