% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingCombinedUserInterface.R
\name{IndexingCombinedUserInterface}
\alias{IndexingCombinedUserInterface}
\title{Indexing combined ui}
\usage{
IndexingCombinedUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing combined ui
}
