% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingSequentialParallel.R
\name{IndexingSequentialParallel}
\alias{IndexingSequentialParallel}
\title{Indexing sequential parallel}
\usage{
IndexingSequentialParallel(
  basename,
  reference,
  gappedIndex,
  indexSplit,
  memory,
  TH_subread
)
}
\arguments{
\item{basename}{output basename}

\item{reference}{reference genome}

\item{gappedIndex}{gapped structure}

\item{indexSplit}{split structure}

\item{memory}{handling memory}

\item{TH_subread}{threshold memory usage}
}
\description{
Indexing sequential parallel
}
