% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingSequentialProgressive.R
\name{IndexingSequentialProgressive}
\alias{IndexingSequentialProgressive}
\title{Indexing sequential progressive}
\usage{
IndexingSequentialProgressive(
  outfolder1,
  outfolder2,
  refgen1,
  refgen2,
  gappedIndex,
  indexSplit,
  memory,
  TH_subread
)
}
\arguments{
\item{outfolder1}{first output folder}

\item{outfolder2}{second output folder}

\item{refgen1}{first reference genome}

\item{refgen2}{second reference genome}

\item{gappedIndex}{gapped structure}

\item{indexSplit}{split structure}

\item{memory}{handling memory}

\item{TH_subread}{threshold memory usage}
}
\description{
Indexing sequential progressive
}
