% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingSequentialServerLogic.R
\name{IndexingSequentialServerLogic}
\alias{IndexingSequentialServerLogic}
\title{Indexing sequential server logic}
\usage{
IndexingSequentialServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing sequential server logic
}
