% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingSequentialUserInterface.R
\name{IndexingSequentialUserInterface}
\alias{IndexingSequentialUserInterface}
\title{Indexing sequential ui}
\usage{
IndexingSequentialUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Indexing sequential ui
}
