% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingBulkServerLogic.R
\name{mappingBulkServerLogic}
\alias{mappingBulkServerLogic}
\title{Mapping bulk server logic}
\usage{
mappingBulkServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping bulk server logic
}
