% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingBulkUserInterface.R
\name{mappingBulkUserInterface}
\alias{mappingBulkUserInterface}
\title{Mapping bulk ui}
\usage{
mappingBulkUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping bulk ui
}
