% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingCombinedServerLogic.R
\name{mappingCombinedServerLogic}
\alias{mappingCombinedServerLogic}
\title{Mapping combined server logic}
\usage{
mappingCombinedServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping combined server logic
}
