% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingCombinedUserInterface.R
\name{mappingCombinedUserInterface}
\alias{mappingCombinedUserInterface}
\title{Mapping combined ui}
\usage{
mappingCombinedUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping combined ui
}
