% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingSequentialServerLogic.R
\name{mappingSequentialServerLogic}
\alias{mappingSequentialServerLogic}
\title{Mapping sequential server logic}
\usage{
mappingSequentialServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping sequential server logic
}
