% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingSequentialUserInterface.R
\name{mappingSequentialUserInterface}
\alias{mappingSequentialUserInterface}
\title{Mapping sequential ui}
\usage{
mappingSequentialUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Mapping sequential ui
}
