% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsinfo.R
\name{mdsinfo}
\alias{mdsinfo}
\title{mdsinfo}
\usage{
mdsinfo(matrix, top, gene.selection)
}
\arguments{
\item{matrix}{A DGEList object.}

\item{top}{Integer. Number of top most variable genes to include in MDS.}

\item{gene.selection}{Method for gene selection: one of "pairwise", "common", or "logFC".}
}
\value{
A list object from  "plotMDS() " containing MDS coordinates and eigenvalues.
}
\description{
Compute MDS coordinates for expression data using limma's plotMDS.
}
\details{
This function performs multidimensional scaling (MDS) on a DGEList or
log-expression matrix using limma's  "plotMDS() " function. It returns the MDS
object containing coordinates and eigenvalues without generating a plot.
}
