% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcainfo.R
\name{pcainfo}
\alias{pcainfo}
\title{pcainfo}
\usage{
pcainfo(logcounts, center, scale)
}
\arguments{
\item{logcounts}{Numeric matrix. Log-CPM values (genes × samples), e.g., from edgeR::cpm..}

\item{center}{Logical. If TRUE, center variables by subtracting the mean (default: TRUE).}

\item{scale}{Logical. If TRUE, scale variables to unit variance (default: FALSE).}
}
\value{
An object of class  "prcomp " containing the PCA results, including loadings,
scores, and explained variance.
}
\description{
Perform Principal Component Analysis (PCA) on log-expression data.
}
\details{
This function transposes a log-count matrix (samples as columns, genes as rows)
and runs PCA using  "stats::prcomp() ", with options to center and scale variables.
}
