% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityCheckAnalysis.R
\name{QualityCheckAnalysis}
\alias{QualityCheckAnalysis}
\title{QUALITY CONTROL ANALYSIS}
\usage{
QualityCheckAnalysis(
  directoryInput,
  inputFormat,
  Nodes,
  ReadsNumber,
  directoryOutput,
  tempFolder
)
}
\arguments{
\item{directoryInput}{sample directory}

\item{inputFormat}{raw read format}

\item{Nodes}{cores}

\item{ReadsNumber}{chunk}

\item{directoryOutput}{output folder}

\item{tempFolder}{temporary folder}
}
\description{
QUALITY CONTROL ANALYSIS
}
