% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlServerLogic.R
\name{qualityControlServerLogic}
\alias{qualityControlServerLogic}
\title{Quality control server logic}
\usage{
qualityControlServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Quality control server logic
}
