% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControlUserInterface.R
\name{qualityControlUserInterface}
\alias{qualityControlUserInterface}
\title{Quality control ui}
\usage{
qualityControlUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Quality control ui
}
