% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceLengthDistributionPlot.R
\name{SequenceLengthDistributionPlot}
\alias{SequenceLengthDistributionPlot}
\title{SequenceLengthDistributionPlot}
\usage{
SequenceLengthDistributionPlot(input_data)
}
\arguments{
\item{input_data}{result tables folder}
}
\description{
SequenceLengthDistributionPlot
}
