% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceLengthDistributionPlotly.R
\name{SequenceLengthDistributionPlotly}
\alias{SequenceLengthDistributionPlotly}
\title{interactive SequenceLengthDistributionPlot}
\usage{
SequenceLengthDistributionPlotly(input_data)
}
\arguments{
\item{input_data}{result tables folder}
}
\description{
interactive SequenceLengthDistributionPlot
}
