% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialAlignment.R
\name{SequentialAlignment}
\alias{SequentialAlignment}
\title{Sequential alignment function}
\usage{
SequentialAlignment(
  lalista,
  nodes,
  readsPath,
  GenomeFirstIndex,
  GenomeSecondIndex,
  outBam1,
  outBam2,
  threads,
  outFormat,
  phredScore,
  maxExtractedSubreads,
  consensusVote,
  mismatchMax,
  uniqueOnly,
  maxMultiMapped,
  indelLength,
  fragmentMinLength,
  fragmentMaxLength,
  matesOrientation,
  readOrderConserved,
  coordinatesSorting,
  allJunctions,
  tempfolder,
  readsAlignedBlock
)
}
\arguments{
\item{lalista}{list of samples}

\item{nodes}{logic cores}

\item{readsPath}{sample folders}

\item{GenomeFirstIndex}{first genome index}

\item{GenomeSecondIndex}{second genome index}

\item{outBam1}{first output folder}

\item{outBam2}{second output folder}

\item{threads}{processes}

\item{outFormat}{BAM or SAM}

\item{phredScore}{quality score}

\item{maxExtractedSubreads}{number of subreads}

\item{consensusVote}{consensus}

\item{mismatchMax}{mismatch}

\item{uniqueOnly}{no multimapping}

\item{maxMultiMapped}{multimapping}

\item{indelLength}{indel}

\item{fragmentMinLength}{fragment minumum length}

\item{fragmentMaxLength}{fragment maximum length}

\item{matesOrientation}{mate orientation}

\item{readOrderConserved}{read order}

\item{coordinatesSorting}{sorting}

\item{allJunctions}{junctions}

\item{tempfolder}{temporary folder}

\item{readsAlignedBlock}{chunks}
}
\description{
Sequential alignment function
}
