% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizationServerLogic.R
\name{SummarizationServerLogic}
\alias{SummarizationServerLogic}
\title{Server function for Summarization module in Shiny application}
\usage{
SummarizationServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Server function for Summarization module in Shiny application
}
