% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizationUserInterface.R
\name{SummarizationUserInterface}
\alias{SummarizationUserInterface}
\title{UI function for Summarization module in Shiny application}
\usage{
SummarizationUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
UI function for Summarization module in Shiny application
}
