% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowServerLogic.R
\name{WorkflowServerLogic}
\alias{WorkflowServerLogic}
\title{Server function for workflow module in Shiny application}
\usage{
WorkflowServerLogic(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
Server function for workflow module in Shiny application
}
