% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowUserInterface.R
\name{WorkflowUserInterface}
\alias{WorkflowUserInterface}
\title{UI function for workflow module in Shiny application}
\usage{
WorkflowUserInterface(id)
}
\arguments{
\item{id}{Shiny module identifier}
}
\description{
UI function for workflow module in Shiny application
}
